+++ RELEASE_NOTES.md
@@ -348,6 +348,7 @@
 
 - Added YAML reading function `iot_data_from_yaml`
 - Added component stopping callback support (called by container before components stopped)
+- Added component starting callback support (called by container before components started)
 - Support removed for Alpine 3.15 and Ubuntu 18.04
 - Added function `iot_data_binary_from_string` to create binary data from string data
 - Added function `iot_data_map_get_array` to find an array in a map